[vis] keep track of task dependence infomation and update the computation graph Now, there are two solutions
1. ompt-tsan-archer-taskdep.cpp
This one simulates what archer does.
When the ompt-dependence callback happens, we save information in parent task and current task.
When a task is schedule, we acquire dependence data from parent.
When a task is finished, we update dependence data in parent.
TODO: one thing todo is that, in archer, the task_t data is deleted only if all its children are finished
This feature is important. Otherwise, it is possible that when a task starts, its parent data is already deleted.
We need to add this feature to our implementation in ompt-tsan-archer-taskdep.cpp

2. ompt-tsan.cpp
This one let each task keep a local vector called depending_task_nodes.
When the ompt-dependence callback happens, we save information in parent task and current task.
When a task is schedule, we acquire dependence data from its local depending_task_nodes vector.
When a task is finished, we don't need to do anything.
TODO: Right now, the vector in parent is not thread-safe, just a STL vector. Need to use a concurrent vector.
Another issue is that, this version now utilized some data structure from archer, which is not necessary.
We can remove that part and just use our own data strucutre.