[ompt52_dpst] clean up code and handle target,teams in OMPT 1. wrap computation graph related code in the macro GRAPH_MACRO 2. wrap some printf in the macro DEBUG_INFO. I keep some important printf outside of the macro so they are always printed. 3. Add ompt_ta_target and ompt_ta_target, just to make sure they are working. 4. Handle teams synchronization in the computation graph. Based on discussion with developers, we realize the synchronization at the end of teams can be tricky. Given the following pragma:
Three levels of parallel regions will be created

level 1: a single paralle region, requested_parallelsim = teams, this is the parallel region for the two teams
level 2: two parallel regions, each one is the implicit parallel for one teams
level 3: two parallel regions, requested_parallelsim = teams=4, each one is created by the parallel construct inside a team

Now the tricky part comes, what is the synchronization for each parallel region in different levels?

level 3: barrier. This is just a normal parallel region, at the end there is a barrier. We will see sync_region_begin and sync_region_end events.

            condition 1          condition 2
level 2:     barrier              no barrier
level 1:     no barrier           barrier

This may not be 100% true, but it is what we observed. We build our computaton graph based on these two conditions.
The changes are in the ompt_ta_implicit_task function.
When an initial task or an implicit task end, we check if it added to other's pending join.
If not, we check again if we are in a teams region; if yes, we add the task to the parallel region's pending joins.